+++ src/mongo/db/commands/find_cmd.cpp
@@ -218,7 +218,7 @@ public:
         }
 
         // Fill out curop information.
-        int ntoreturn = lpq->getBatchSize().value_or(0);
+        long long ntoreturn = lpq->getBatchSize().value_or(0);
         beginQueryOp(txn, nss, cmdObj, ntoreturn, lpq->getSkip());
 
         // 1b) Finish the parsing step by using the LiteParsedQuery to create a CanonicalQuery.
@@ -264,7 +264,7 @@ public:
         if (!collection) {
             // No collection. Just fill out curop indicating that there were zero results and
             // there is no ClientCursor id, and then return.
-            const int numResults = 0;
+            const long long numResults = 0;
             const CursorId cursorId = 0;
             endQueryOp(txn, *exec, dbProfilingLevel, numResults, cursorId);
             appendCursorResponseObject(cursorId, nss.ns(), BSONArray(), &result);
@@ -300,7 +300,7 @@ public:
         BSONArrayBuilder firstBatch;
         BSONObj obj;
         PlanExecutor::ExecState state;
-        int numResults = 0;
+        long long numResults = 0;
         while (!enoughForFirstBatch(pq, numResults, firstBatch.len()) &&
                PlanExecutor::ADVANCED == (state = cursorExec->getNext(&obj, NULL))) {
             // If adding this object will cause us to exceed the BSON size limit, then we stash
